split errors from single file. added tests.
